Left with handlers that do not require moving to a separate module; added docstring;

+++ handlers/handlers.py
@@ -1,6 +1,10 @@
 """Bot Handlers
 
-pass
+This module contains general-purpose bot handlers that are not directly related
+to a specific feature but support essential functionality.
+
+Currently, it includes:
+    - `update` callback handler
 """
 
 import asyncio
@@ -20,9 +24,16 @@ logger = logging.getLogger(__name__)
 
 @handlers_router.callback_query(F.data == 'update')
 async def update(callback: CallbackQuery) -> None:
-    await callback.answer(handlers_lex['reboot'])
-    if 'updater.exe' in os.getcwd():
-        subprocess.Popen("updater.exe")
+    """Checks if `updater.exe` exists in the bot's directory.
+       If found, notifies the user, starts the updater, and stops the bot gracefully.
+       Otherwise, sends an error message.
+    """
+    bot_dir = os.path.dirname(os.path.abspath(__file__))
+    updater_filename = "updater.exe"
+    updater_path = os.path.join(bot_dir, updater_filename)
+    if os.path.exists(updater_path):
+        await callback.answer(handlers_lex['reboot'])
+        subprocess.Popen([updater_path], close_fds=True)
         await config.dp.stop_polling()
         loop = asyncio.get_running_loop()
         loop.stop()
